+++ controller/lib/src/adp.h
@@ -39,8 +39,8 @@ namespace avdecc_lib
         struct jdksavdecc_frame cmd_frame; // Structure containing the Ethernet Frame fields
         struct jdksavdecc_adpdu adpdu; // Structure containing the ADPDU fields
         uint8_t *adp_frame; // Point to a raw memory buffer to read from
-        int frame_read_returned; // Status of extracting Ethernet Frame information from a network buffer
-        int adpdu_read_returned; // Status of extracting ADPDU information from a network buffer
+        ssize_t frame_read_returned; // Status of extracting Ethernet Frame information from a network buffer
+        ssize_t adpdu_read_returned; // Status of extracting ADPDU information from a network buffer
 
     public:
         /**
