+++ controller/lib/src/configuration_descriptor_imp.cpp
@@ -38,7 +38,7 @@
 
 namespace avdecc_lib
 {
-    configuration_descriptor_imp::configuration_descriptor_imp(end_station_imp *end_station_obj, const uint8_t *frame, size_t pos, size_t frame_len) : descriptor_base_imp(end_station_obj)
+    configuration_descriptor_imp::configuration_descriptor_imp(end_station_imp *end_station_obj, const uint8_t *frame, ssize_t pos, size_t frame_len) : descriptor_base_imp(end_station_obj)
     {
         config_desc_read_returned = jdksavdecc_descriptor_configuration_read(&config_desc, frame, pos, frame_len);
 
@@ -190,158 +190,158 @@ namespace avdecc_lib
         }
     }
 
-    void configuration_descriptor_imp::store_audio_unit_desc(end_station_imp *end_station_obj, const uint8_t *frame, size_t pos, size_t frame_len)
+    void configuration_descriptor_imp::store_audio_unit_desc(end_station_imp *end_station_obj, const uint8_t *frame, ssize_t pos, size_t frame_len)
     {
         audio_unit_descriptor_imp *d = new audio_unit_descriptor_imp(end_station_obj, frame, pos, frame_len);
         add_or_replace_descriptor_and_sort(d, audio_unit_desc_vec);
     }
 
-    void configuration_descriptor_imp::store_stream_input_desc(end_station_imp *end_station_obj, const uint8_t *frame, size_t pos, size_t frame_len)
+    void configuration_descriptor_imp::store_stream_input_desc(end_station_imp *end_station_obj, const uint8_t *frame, ssize_t pos, size_t frame_len)
     {
         stream_input_descriptor_imp *d = new stream_input_descriptor_imp(end_station_obj, frame, pos, frame_len);
         add_or_replace_descriptor_and_sort(d, stream_input_desc_vec);
     }
 
-    void configuration_descriptor_imp::store_stream_output_desc(end_station_imp *end_station_obj, const uint8_t *frame, size_t pos, size_t frame_len)
+    void configuration_descriptor_imp::store_stream_output_desc(end_station_imp *end_station_obj, const uint8_t *frame, ssize_t pos, size_t frame_len)
     {
         stream_output_descriptor_imp *d = new stream_output_descriptor_imp(end_station_obj, frame, pos, frame_len);
         add_or_replace_descriptor_and_sort(d, stream_output_desc_vec);
     }
 
-    void configuration_descriptor_imp::store_jack_input_desc(end_station_imp *end_station_obj, const uint8_t *frame, size_t pos, size_t frame_len)
+    void configuration_descriptor_imp::store_jack_input_desc(end_station_imp *end_station_obj, const uint8_t *frame, ssize_t pos, size_t frame_len)
     {
         jack_input_descriptor_imp *d = new jack_input_descriptor_imp(end_station_obj, frame, pos, frame_len);
         add_or_replace_descriptor_and_sort(d, jack_input_desc_vec);
     }
 
-    void configuration_descriptor_imp::store_jack_output_desc(end_station_imp *end_station_obj, const uint8_t *frame, size_t pos, size_t frame_len)
+    void configuration_descriptor_imp::store_jack_output_desc(end_station_imp *end_station_obj, const uint8_t *frame, ssize_t pos, size_t frame_len)
     {
         jack_output_descriptor_imp *d = new jack_output_descriptor_imp(end_station_obj, frame, pos, frame_len);
         add_or_replace_descriptor_and_sort(d, jack_output_desc_vec);
     }
 
-    void configuration_descriptor_imp::store_avb_interface_desc(end_station_imp *end_station_obj, const uint8_t *frame, size_t pos, size_t frame_len)
+    void configuration_descriptor_imp::store_avb_interface_desc(end_station_imp *end_station_obj, const uint8_t *frame, ssize_t pos, size_t frame_len)
     {
         avb_interface_descriptor_imp *d = new avb_interface_descriptor_imp(end_station_obj, frame, pos, frame_len);
         add_or_replace_descriptor_and_sort(d, avb_interface_desc_vec);
     }
 
-    void configuration_descriptor_imp::store_clock_source_desc(end_station_imp *end_station_obj, const uint8_t *frame, size_t pos, size_t frame_len)
+    void configuration_descriptor_imp::store_clock_source_desc(end_station_imp *end_station_obj, const uint8_t *frame, ssize_t pos, size_t frame_len)
     {
         clock_source_descriptor_imp *d = new clock_source_descriptor_imp(end_station_obj, frame, pos, frame_len);
         add_or_replace_descriptor_and_sort(d, clock_source_desc_vec);
     }
 
-    void configuration_descriptor_imp::store_locale_desc(end_station_imp *end_station_obj, const uint8_t *frame, size_t pos, size_t frame_len)
+    void configuration_descriptor_imp::store_locale_desc(end_station_imp *end_station_obj, const uint8_t *frame, ssize_t pos, size_t frame_len)
     {
         locale_descriptor_imp *d = new locale_descriptor_imp(end_station_obj, frame, pos, frame_len);
         add_or_replace_descriptor_and_sort(d, locale_desc_vec);
     }
 
-    void configuration_descriptor_imp::store_strings_desc(end_station_imp *end_station_obj, const uint8_t *frame, size_t pos, size_t frame_len)
+    void configuration_descriptor_imp::store_strings_desc(end_station_imp *end_station_obj, const uint8_t *frame, ssize_t pos, size_t frame_len)
     {
         strings_descriptor_imp *d = new strings_descriptor_imp(end_station_obj, frame, pos, frame_len);
         add_or_replace_descriptor_and_sort(d, strings_desc_vec);
     }
 
-    void configuration_descriptor_imp::store_stream_port_input_desc(end_station_imp *end_station_obj, const uint8_t *frame, size_t pos, size_t frame_len)
+    void configuration_descriptor_imp::store_stream_port_input_desc(end_station_imp *end_station_obj, const uint8_t *frame, ssize_t pos, size_t frame_len)
     {
         stream_port_input_desc_vec.push_back(new stream_port_input_descriptor_imp(end_station_obj, frame, pos, frame_len));
     }
 
-    void configuration_descriptor_imp::store_stream_port_output_desc(end_station_imp *end_station_obj, const uint8_t *frame, size_t pos, size_t frame_len)
+    void configuration_descriptor_imp::store_stream_port_output_desc(end_station_imp *end_station_obj, const uint8_t *frame, ssize_t pos, size_t frame_len)
     {
         stream_port_output_desc_vec.push_back(new stream_port_output_descriptor_imp(end_station_obj, frame, pos, frame_len));
     }
 
-    void configuration_descriptor_imp::store_audio_cluster_desc(end_station_imp *end_station_obj, const uint8_t *frame, size_t pos, size_t frame_len)
+    void configuration_descriptor_imp::store_audio_cluster_desc(end_station_imp *end_station_obj, const uint8_t *frame, ssize_t pos, size_t frame_len)
     {
         audio_cluster_descriptor_imp *d = new audio_cluster_descriptor_imp(end_station_obj, frame, pos, frame_len);
         add_or_replace_descriptor_and_sort(d, audio_cluster_desc_vec);
     }
 
-    void configuration_descriptor_imp::store_audio_map_desc(end_station_imp *end_station_obj, const uint8_t *frame, size_t pos, size_t frame_len)
+    void configuration_descriptor_imp::store_audio_map_desc(end_station_imp *end_station_obj, const uint8_t *frame, ssize_t pos, size_t frame_len)
     {
         audio_map_descriptor_imp *d = new audio_map_descriptor_imp(end_station_obj, frame, pos, frame_len);
         add_or_replace_descriptor_and_sort(d, audio_map_desc_vec);
     }
 
-    void configuration_descriptor_imp::store_clock_domain_desc(end_station_imp *end_station_obj, const uint8_t *frame, size_t pos, size_t frame_len)
+    void configuration_descriptor_imp::store_clock_domain_desc(end_station_imp *end_station_obj, const uint8_t *frame, ssize_t pos, size_t frame_len)
     {
         clock_domain_descriptor_imp *d = new clock_domain_descriptor_imp(end_station_obj, frame, pos, frame_len);
         add_or_replace_descriptor_and_sort(d, clock_domain_desc_vec);
     }
 
-    uint32_t STDCALL configuration_descriptor_imp::audio_unit_desc_count()
+    size_t STDCALL configuration_descriptor_imp::audio_unit_desc_count()
     {
         return audio_unit_desc_vec.size();
     }
 
-    uint32_t STDCALL configuration_descriptor_imp::stream_input_desc_count()
+    size_t STDCALL configuration_descriptor_imp::stream_input_desc_count()
     {
         return stream_input_desc_vec.size();
     }
-    uint32_t STDCALL configuration_descriptor_imp::stream_output_desc_count()
+    size_t STDCALL configuration_descriptor_imp::stream_output_desc_count()
     {
         return stream_output_desc_vec.size();
     }
 
-    uint32_t STDCALL configuration_descriptor_imp::jack_input_desc_count()
+    size_t STDCALL configuration_descriptor_imp::jack_input_desc_count()
     {
         return jack_input_desc_vec.size();
     }
 
-    uint32_t STDCALL configuration_descriptor_imp::jack_output_desc_count()
+    size_t STDCALL configuration_descriptor_imp::jack_output_desc_count()
     {
         return jack_output_desc_vec.size();
     }
 
-    uint32_t STDCALL configuration_descriptor_imp::avb_interface_desc_count()
+    size_t STDCALL configuration_descriptor_imp::avb_interface_desc_count()
     {
         return avb_interface_desc_vec.size();
     }
 
-    uint32_t STDCALL configuration_descriptor_imp::clock_source_desc_count()
+    size_t STDCALL configuration_descriptor_imp::clock_source_desc_count()
     {
         return clock_source_desc_vec.size();
     }
 
-    uint32_t STDCALL configuration_descriptor_imp::locale_desc_count()
+    size_t STDCALL configuration_descriptor_imp::locale_desc_count()
     {
         return locale_desc_vec.size();
     }
 
-    uint32_t STDCALL configuration_descriptor_imp::strings_desc_count()
+    size_t STDCALL configuration_descriptor_imp::strings_desc_count()
     {
         return strings_desc_vec.size();
     }
 
-    uint32_t STDCALL configuration_descriptor_imp::stream_port_input_desc_count()
+    size_t STDCALL configuration_descriptor_imp::stream_port_input_desc_count()
     {
         return stream_port_input_desc_vec.size();
     }
 
-    uint32_t STDCALL configuration_descriptor_imp::stream_port_output_desc_count()
+    size_t STDCALL configuration_descriptor_imp::stream_port_output_desc_count()
     {
         return stream_port_output_desc_vec.size();
     }
 
-    uint32_t STDCALL configuration_descriptor_imp::audio_cluster_desc_count()
+    size_t STDCALL configuration_descriptor_imp::audio_cluster_desc_count()
     {
         return audio_cluster_desc_vec.size();
     }
 
-    uint32_t STDCALL configuration_descriptor_imp::audio_map_desc_count()
+    size_t STDCALL configuration_descriptor_imp::audio_map_desc_count()
     {
         return audio_map_desc_vec.size();
     }
 
-    uint32_t STDCALL configuration_descriptor_imp::clock_domain_desc_count()
+    size_t STDCALL configuration_descriptor_imp::clock_domain_desc_count()
     {
         return clock_domain_desc_vec.size();
     }
 
-    audio_unit_descriptor * STDCALL configuration_descriptor_imp::get_audio_unit_desc_by_index(uint32_t audio_unit_desc_index)
+    audio_unit_descriptor * STDCALL configuration_descriptor_imp::get_audio_unit_desc_by_index(size_t audio_unit_desc_index)
     {
         bool is_valid = (audio_unit_desc_index < audio_unit_desc_vec.size());
 
@@ -357,7 +357,7 @@ namespace avdecc_lib
         return NULL;
     }
 
-    stream_input_descriptor * STDCALL configuration_descriptor_imp::get_stream_input_desc_by_index(uint32_t stream_input_desc_index)
+    stream_input_descriptor * STDCALL configuration_descriptor_imp::get_stream_input_desc_by_index(size_t stream_input_desc_index)
     {
         bool is_valid = (stream_input_desc_index < stream_input_desc_vec.size());
 
@@ -373,7 +373,7 @@ namespace avdecc_lib
         return NULL;
     }
 
-    stream_output_descriptor * STDCALL configuration_descriptor_imp::get_stream_output_desc_by_index(uint32_t stream_output_desc_index)
+    stream_output_descriptor * STDCALL configuration_descriptor_imp::get_stream_output_desc_by_index(size_t stream_output_desc_index)
     {
         bool is_valid = (stream_output_desc_index < stream_output_desc_vec.size());
 
@@ -389,7 +389,7 @@ namespace avdecc_lib
         return NULL;
     }
 
-    jack_input_descriptor * STDCALL configuration_descriptor_imp::get_jack_input_desc_by_index(uint32_t jack_input_desc_index)
+    jack_input_descriptor * STDCALL configuration_descriptor_imp::get_jack_input_desc_by_index(size_t jack_input_desc_index)
     {
         bool is_valid = (jack_input_desc_index < jack_input_desc_vec.size());
 
@@ -405,7 +405,7 @@ namespace avdecc_lib
         return NULL;
     }
 
-    jack_output_descriptor * STDCALL configuration_descriptor_imp::get_jack_output_desc_by_index(uint32_t jack_output_desc_index)
+    jack_output_descriptor * STDCALL configuration_descriptor_imp::get_jack_output_desc_by_index(size_t jack_output_desc_index)
     {
         bool is_valid = (jack_output_desc_index < jack_output_desc_vec.size());
 
@@ -421,7 +421,7 @@ namespace avdecc_lib
         return NULL;
     }
 
-    avb_interface_descriptor * STDCALL configuration_descriptor_imp::get_avb_interface_desc_by_index(uint32_t avb_interface_desc_index)
+    avb_interface_descriptor * STDCALL configuration_descriptor_imp::get_avb_interface_desc_by_index(size_t avb_interface_desc_index)
     {
         bool is_valid = (avb_interface_desc_index < avb_interface_desc_vec.size());
 
@@ -437,7 +437,7 @@ namespace avdecc_lib
         return NULL;
     }
 
-    clock_source_descriptor * STDCALL configuration_descriptor_imp::get_clock_source_desc_by_index(uint32_t clock_source_desc_index)
+    clock_source_descriptor * STDCALL configuration_descriptor_imp::get_clock_source_desc_by_index(size_t clock_source_desc_index)
     {
         bool is_valid = (clock_source_desc_index < clock_source_desc_vec.size());
 
@@ -453,7 +453,7 @@ namespace avdecc_lib
         return NULL;
     }
 
-    locale_descriptor * STDCALL configuration_descriptor_imp::get_locale_desc_by_index(uint32_t locale_desc_index)
+    locale_descriptor * STDCALL configuration_descriptor_imp::get_locale_desc_by_index(size_t locale_desc_index)
     {
         bool is_valid = (locale_desc_index < locale_desc_vec.size());
 
@@ -469,7 +469,7 @@ namespace avdecc_lib
         return NULL;
     }
 
-    strings_descriptor * STDCALL configuration_descriptor_imp::get_strings_desc_by_index(uint32_t strings_desc_index)
+    strings_descriptor * STDCALL configuration_descriptor_imp::get_strings_desc_by_index(size_t strings_desc_index)
     {
         bool is_valid = (strings_desc_index < strings_desc_vec.size());
 
@@ -485,7 +485,7 @@ namespace avdecc_lib
         return NULL;
     }
 
-    stream_port_input_descriptor * STDCALL configuration_descriptor_imp::get_stream_port_input_desc_by_index(uint32_t stream_port_input_desc_index)
+    stream_port_input_descriptor * STDCALL configuration_descriptor_imp::get_stream_port_input_desc_by_index(size_t stream_port_input_desc_index)
     {
         bool is_valid = (stream_port_input_desc_index < stream_port_input_desc_vec.size());
 
@@ -501,7 +501,7 @@ namespace avdecc_lib
         return NULL;
     }
 
-    stream_port_output_descriptor * STDCALL configuration_descriptor_imp::get_stream_port_output_desc_by_index(uint32_t stream_port_output_desc_index)
+    stream_port_output_descriptor * STDCALL configuration_descriptor_imp::get_stream_port_output_desc_by_index(size_t stream_port_output_desc_index)
     {
         bool is_valid = (stream_port_output_desc_index < stream_port_output_desc_vec.size());
 
@@ -517,7 +517,7 @@ namespace avdecc_lib
         return NULL;
     }
 
-    audio_cluster_descriptor * STDCALL configuration_descriptor_imp::get_audio_cluster_desc_by_index(uint32_t audio_cluster_desc_index)
+    audio_cluster_descriptor * STDCALL configuration_descriptor_imp::get_audio_cluster_desc_by_index(size_t audio_cluster_desc_index)
     {
         bool is_valid = (audio_cluster_desc_index < audio_cluster_desc_vec.size());
 
@@ -533,7 +533,7 @@ namespace avdecc_lib
         return NULL;
     }
 
-    audio_map_descriptor * STDCALL configuration_descriptor_imp::get_audio_map_desc_by_index(uint32_t audio_map_desc_index)
+    audio_map_descriptor * STDCALL configuration_descriptor_imp::get_audio_map_desc_by_index(size_t audio_map_desc_index)
     {
         bool is_valid = (audio_map_desc_index < audio_map_desc_vec.size());
 
@@ -549,7 +549,7 @@ namespace avdecc_lib
         return NULL;
     }
 
-    clock_domain_descriptor * STDCALL configuration_descriptor_imp::get_clock_domain_desc_by_index(uint32_t clock_domain_desc_index)
+    clock_domain_descriptor * STDCALL configuration_descriptor_imp::get_clock_domain_desc_by_index(size_t clock_domain_desc_index)
     {
         bool is_valid = (clock_domain_desc_index < clock_domain_desc_vec.size());
